feat: rework signatories panel with log tab and header actions
