+++ src/renderer/features/multisig-operations/components/OperationSignatories.tsx
@@ -1,5 +1,5 @@
 import { useUnit } from 'effector-react';
-import { type ReactNode, useCallback, useMemo, useState } from 'react';
+import { type ReactNode, useMemo, useState } from 'react';
 
 import {
   type Chain,
@@ -13,10 +13,11 @@ import {
 } from '@/shared/core';
 import { Slot, createSlot } from '@/shared/di';
 import { useI18n } from '@/shared/i18n';
-import { nonNullable, toAddress } from '@/shared/lib/utils';
+import { cnTw, nonNullable, toAddress } from '@/shared/lib/utils';
 import { type AccountId } from '@/shared/polkadotjs-schemas';
-import { Button, CaptionText, FootnoteText, Icon, SmallTitleText } from '@/shared/ui';
+import { CountChip, FootnoteText, IconButton, SmallTitleText } from '@/shared/ui';
 import { Address } from '@/shared/ui-entities';
+import { Copy, Tooltip } from '@/shared/ui-kit';
 import {
   type AnyAccount,
   type MultisigOperation,
@@ -31,7 +32,7 @@ import { SignatoryCard } from '@/entities/signatory';
 import { accountUtils, walletModel } from '@/entities/wallet';
 import { NamedAccount } from '@/widgets/NameResolver';
 
-import LogModal from './LogModal';
+import { OperationLog } from './OperationLog';
 
 type SignatoryAddressProps = {
   accountId: AccountId;
@@ -70,17 +71,19 @@ type WalletSignatory = Signatory & { account: AnyAccount; wallet: Wallet };
 type Props = {
   operation: MultisigOperation;
   account: MultisigAccount | FlexibleMultisigAccount;
+  deepLink: string;
 };
 
-export const OperationSignatories = ({ operation, account }: Props) => {
+type ActiveTab = 'signatories' | 'log';
+
+export const OperationSignatories = ({ operation, account, deepLink }: Props) => {
   const { t } = useI18n();
   const chain = useChain(operation.chainId);
 
   const wallets = useUnit(walletModel.$wallets);
   const accountsList = useUnit(accounts.$list);
 
-  const [isLogModalOpen, setIsLogModalOpen] = useState(false);
-  const closeLogModal = useCallback(() => setIsLogModalOpen(false), []);
+  const [activeTab, setActiveTab] = useState<ActiveTab>('signatories');
 
   const approvals = multisigOperationService.getApprovals(operation);
   const cancellation = operation.events.filter(e => e.status === 'reject');
@@ -115,8 +118,7 @@ export const OperationSignatories = ({ operation, account }: Props) => {
     return acc;
   }, []);
 
-  const walletSignatoriesIds = walletSignatories.map(a => a.accountId);
-  const contactSignatories = account.signatories.filter(s => !walletSignatoriesIds.includes(s.accountId));
+  const walletSignatoriesMap = new Map(walletSignatories.map(signatory => [signatory.accountId, signatory]));
 
   // Contact-backed external multisigs aren't part of the user's account
   // graph, so the "Open overview" structure view has nothing meaningful to
@@ -150,24 +152,29 @@ export const OperationSignatories = ({ operation, account }: Props) => {
 
   return (
     <div className="flex flex-col border-r border-divider p-4">
-      <div className="mb-4 flex items-center justify-between">
-        <div className="flex items-center gap-2">
-          <SmallTitleText>{t('operation.signatoriesTitle')}</SmallTitleText>
-          <Button
-            pallet="secondary"
-            variant="fill"
-            size="sm"
-            prefixElement={<Icon name="chat" size={16} />}
-            suffixElement={
-              <CaptionText className="rounded-full bg-chip-icon px-1.5 pt-px pb-[2px] text-white!">
-                {operation.events.length}
-              </CaptionText>
-            }
-            onClick={() => setIsLogModalOpen(true)}
-          >
+      <div className="mb-4 flex items-center gap-2">
+        <button type="button" onClick={() => setActiveTab('signatories')}>
+          <SmallTitleText className={cnTw(activeTab !== 'signatories' && 'text-text-tertiary')}>
+            {t('operation.signatoriesTitle')}
+          </SmallTitleText>
+        </button>
+
+        <button
+          type="button"
+          className={cnTw(
+            'flex items-center gap-1.5 rounded-md px-2 py-1 transition-colors',
+            activeTab === 'log' ? 'bg-tab-background' : 'hover:bg-action-background-hover',
+          )}
+          onClick={() => setActiveTab('log')}
+        >
+          <FootnoteText className={cnTw(activeTab !== 'log' && 'text-text-tertiary')}>
             {t('operation.logButton')}
-          </Button>
-        </div>
+          </FootnoteText>
+          <CountChip count={operation.events.length} />
+        </button>
+
+        <span className="flex-1" />
+
         {!isExternalMultisig && (
           <Slot
             id={operationOverviewSlot}
@@ -175,62 +182,62 @@ export const OperationSignatories = ({ operation, account }: Props) => {
               walletAccounts: overviewAccounts,
               initialChainId: operation.chainId,
               exclusive: true,
+              // Wrapped in a plain div: Modal.Trigger clones props onto its child via
+              // Radix asChild, and our Tooltip component doesn't forward unknown props,
+              // so the click-to-open handler must land on a real DOM node.
               trigger: (
-                <Button pallet="primary" variant="text" size="sm">
-                  {t('operation.openOverviewButton')}
-                </Button>
+                <div>
+                  <Tooltip>
+                    <Tooltip.Trigger>
+                      <IconButton name="accountStructure" className="text-icon-default" />
+                    </Tooltip.Trigger>
+                    <Tooltip.Content>{t('operation.openOverviewButton')}</Tooltip.Content>
+                  </Tooltip>
+                </div>
               ),
             }}
           />
         )}
+
+        <Tooltip>
+          <Tooltip.Trigger>
+            <Copy value={deepLink} notification={t('general.notifications.operationLinkCopied')}>
+              <IconButton name="share" className="text-icon-default" />
+            </Copy>
+          </Tooltip.Trigger>
+          <Tooltip.Content>{t('operations.shareOperationTooltip')}</Tooltip.Content>
+        </Tooltip>
       </div>
 
-      <div className="flex flex-col gap-y-2">
-        {Boolean(walletSignatories.length) && (
-          <div>
-            <FootnoteText className="text-text-tertiary" as="h4">
-              {t('operation.walletSignatoriesTitle')}
-            </FootnoteText>
-            <ul className="flex flex-col">
-              {walletSignatories.map(signatory => (
-                <SignatoryCard
-                  key={signatory.accountId}
-                  status={operationDetailsUtils.getSignatoryStatus(operation.events, signatory.accountId)}
-                >
+      {activeTab === 'signatories' ? (
+        <ul className="flex flex-col gap-y-2">
+          {signatoriesList.map(signatory => {
+            const walletSignatory = walletSignatoriesMap.get(signatory.accountId);
+
+            return (
+              <SignatoryCard
+                key={signatory.accountId}
+                status={operationDetailsUtils.getSignatoryStatus(operation.events, signatory.accountId)}
+              >
+                {walletSignatory ? (
                   <NamedAccount
-                    accountId={signatory.account.accountId}
+                    accountId={walletSignatory.account.accountId}
                     chain={chain ?? undefined}
-                    title={signatory.account.name}
-                    wallet={signatory.wallet}
+                    title={walletSignatory.account.name}
+                    wallet={walletSignatory.wallet}
                     variant="short"
                     iconSize={20}
                   />
-                </SignatoryCard>
-              ))}
-            </ul>
-          </div>
-        )}
-
-        {Boolean(contactSignatories.length) && (
-          <div>
-            <FootnoteText className="text-text-tertiary" as="h4">
-              {t('operation.contactSignatoriesTitle')}
-            </FootnoteText>
-            <ul className="flex flex-col">
-              {contactSignatories.map(signatory => (
-                <SignatoryCard
-                  key={signatory.accountId}
-                  status={operationDetailsUtils.getSignatoryStatus(operation.events, signatory.accountId)}
-                >
+                ) : (
                   <SignatoryAddress accountId={signatory.accountId} chain={chain} />
-                </SignatoryCard>
-              ))}
-            </ul>
-          </div>
-        )}
-      </div>
-
-      <LogModal isOpen={isLogModalOpen} operation={operation} account={account} chain={chain} onClose={closeLogModal} />
+                )}
+              </SignatoryCard>
+            );
+          })}
+        </ul>
+      ) : (
+        <OperationLog operation={operation} account={account} chain={chain} />
+      )}
     </div>
   );
 };
