+++ src/renderer/features/multisig-operations/index.ts
@@ -1,4 +1,4 @@
-export { operationLogTitleTransformer } from './components/LogModal';
+export { operationLogTitleTransformer } from './components/OperationLog';
 export { type OperationTitle, operationTitleTransformer } from './components/Operation';
 export { operationDetailsSlot } from './components/OperationFullInfo';
 export { operationOverviewSlot } from './components/OperationSignatories';
